CFI: Refactor: Split into 3 modules
